Split negate into its own scalar.
