+++ rust/experimental/query_engine/kql-parser/src/lib.rs
@@ -9,6 +9,7 @@ pub(crate) mod query_expression;
 pub(crate) mod scalar_conditional_function_expressions;
 pub(crate) mod scalar_conversion_function_expressions;
 pub(crate) mod scalar_expression;
+pub(crate) mod scalar_mathematical_function_expressions;
 pub(crate) mod scalar_primitive_expressions;
 pub(crate) mod scalar_string_function_expressions;
 pub(crate) mod shared_expressions;
